+++ CHANGELOG.rst
@@ -9,8 +9,11 @@ Changelog
   - keep the format consistent (88 char width, Y/M/D date format) and do not use tabs but
     use spaces for formatting
 
-.. Unreleased
-.. ----------
+Unreleased
+----------
+Philip Loche
+
+- Add united atom elements to electron density calculations (#509)
 
 v0.11.1 (2025/07/04)
 --------------------